% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mz.unity.R
\name{aggregate.children}
\alias{aggregate.children}
\title{Transverse a directed graph collecting all child nodes.}
\usage{
\method{aggregate}{children}(rels, node)
}
\arguments{
\item{rels}{Matrix. Representation of the relationship graph with column names "A" and "B.*".}

\item{node}{Integer. The node at which to start looking for children.}
}
\value{
Numeric. A vector containing all child nodes.
}
\description{
\code{aggregate.children} returns a vector of nodes which are children at all depths of a specified node.  This is convenient for gathering isotopic spectra.
}

