% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mz.unity.R
\name{aggregate.self}
\alias{aggregate.self}
\title{Transverse a graph collecting all connected nodes.}
\usage{
\method{aggregate}{self}(rels, force = F)
}
\arguments{
\item{rels}{Matrix. Representation of the relationship graph with column names "A" and "B.*".}
}
\value{
Numeric. A matrix containing group assignments for the correspond columns in "A" and "B.*"
}
\description{
\code{aggregate.self} returns a vector of group assignments containing nodes whose parents already belong to the group.  This is convenient for gathering compound groups.
}

