% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.functions.R
\name{expandGraph}
\alias{expandGraph}
\title{Turns a mz.unity relationship output into a graph structured data.frame.}
\usage{
expandGraph(mat, cols.from = "^B$|^B.", cols.to = "^A$|^A.")
}
\arguments{
\item{mat}{Matrix output from mz.unity}

\item{cols.from}{Column names to use as edge sources, expressed as a regex.}

\item{cols.to}{Column names to use as edge targets, expressed as a regex.}
}
\description{
Turns a mz.unity relationship output into a graph structured data.frame.
}
