% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granular.formula.R
\docType{data}
\name{M.iso}
\alias{M.iso}
\title{Common isotope differences.}
\format{
An object of class \code{data.frame} with 10 rows and 6 columns.
}
\usage{
M.iso
}
\description{
Common isotope differences.
}
\keyword{datasets}
