% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granular.formula.R
\docType{data}
\name{M.z}
\alias{M.z}
\title{Name, charge added, mass of adduct}
\format{
An object of class \code{data.frame} with 5 rows and 5 columns.
}
\usage{
M.z
}
\description{
Name, charge added, mass of adduct
}
\keyword{datasets}
