% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.functions.R
\name{reindex}
\alias{reindex}
\title{Takes a matrix and reindexes cells reminescent of match()}
\usage{
reindex(pConv, col.regex, indices)
}
\arguments{
\item{pConv}{Matrix. With integer columns referenced in col.regex to be reindexed.}

\item{col.regex}{Character. Regex matching to the column names of columns to be reindexed.}

\item{indices}{Numeric. Numeric vector with new values.  \code{indices[pConv[1,1]]} will be the replacement value}
}
\value{
A matrix with referenced cells replaced by new values
}
\description{
\code{reindex} takes a matrix and reindexes cells reminescent of match()
}
\details{
This is useful for performing a search with a subset of granular formula, and reindexing to some master ID index.
}

