% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{unity.specgraph}
\alias{unity.specgraph}
\title{Plot the results of mz-unity.search as a spectrum graph.}
\usage{
unity.specgraph(df, spec, edgetypes, speccolor = "black",
  circlecolor = "forestgreen")
}
\arguments{
\item{df}{data.frame Columns named "to", "from", "rel". "rel" is used for coloring.}

\item{spec}{data.frame Columns named "mz", "int".  Rows correspond to vertex indices "to" and "from" in df.}

\item{edgetypes}{data.frame Defines the colors of each edge rel type. COlumn names "rel", "label", "color"}

\item{speccolor}{character The color for the spectrum plot}

\item{circlecolor}{character The color for the circles at the spectrum apexes.}
}
\value{
A ggplot object contianing the plot.
}
\description{
\code{unity.specgraph} plots a graph of detected relationships on top of a mass spectrum.
}
\seealso{
See \url{https://github.com/nathaniel-mahieu/mz-unity} for examples.
}

