% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.functions.R
\name{expandGraph}
\alias{expandGraph}
\title{Turns a mz.unity relationship output into a graph structured data.frame.}
\usage{
expandGraph(mat, cols.from = "^B$|^B.", cols.to = "^A$|^A.")
}
\description{
Turns a mz.unity relationship output into a graph structured data.frame.
}

